requirements.txt
alabaster==0.7.12
arrow==1.2.1
astroid==2.9.0
atomicwrites==1.4.0
attrs==21.2.0
autopep8==1.6.0
Babel==2.9.1
backcall==0.2.0
bcrypt==3.2.0
binaryornot==0.4.4
black==21.11b1
bleach==4.1.0
brotlipy==0.7.0
certifi==2021.5.30
cffi
chardet
click==8.0.3
cloudpickle==2.0.0
conda==4.10.3
conda-package-handling
cookiecutter==1.7.3
cryptography
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
diff-match-patch==20200713
dill==0.3.4
docutils==0.17.1
entrypoints==0.3
flake8==4.0.1
globus-sdk==3.1.0
gprof2dot==2021.2.21
idna
imagesize==1.3.0
importlib-metadata==4.8.2
inflection==0.5.1
intervaltree==3.1.0
ipykernel==6.5.1
ipython==7.30.0
ipython-genutils==0.2.0
isort==5.10.1
jedi==0.18.1
jeepney==0.7.1
jellyfish==0.8.9
Jinja2==3.0.3
jinja2-time==0.2.0
jsonschema==4.2.1
jupyter-client==7.1.0
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
keyring==23.4.0
lazy-object-proxy==1.6.0
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mccabe==0.6.1
mistune==0.8.4
mypy-extensions==0.4.3
nbclient==0.5.9
nbconvert==6.3.0
nbformat==5.1.3
nest-asyncio==1.5.1
numpy==1.21.4
numpydoc==1.1.0
packaging==21.3
pandocfilters==1.5.0
paramiko==2.8.0
parsl==1.1.0
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
platformdirs==2.4.0
pluggy==1.0.0
poyo==0.5.0
prompt-toolkit==3.0.23
psutil==5.8.0
ptyprocess==0.7.0
pycodestyle==2.8.0
pycosat==0.6.3
pycparser
pydocstyle==6.1.1
pyflakes==2.4.0
Pygments==2.10.0
pyinstrument==4.1.1
PyJWT==2.3.0
pylint==2.12.1
pyls-spyder==0.4.0
PyNaCl==1.4.0
pyOpenSSL
pyparsing==3.0.6
pyprof2calltree==1.4.5
PyQt5==5.12.3
PyQt5-sip==12.9.0
PyQtWebEngine==5.12.1
pyrsistent==0.18.0
PySocks
python-dateutil==2.8.2
python-lsp-black==1.0.0
python-lsp-jsonrpc==1.0.0
python-lsp-server==1.3.2
python-slugify==5.0.2
pytz==2021.3
pyxdg==0.27
pyzmq==22.3.0
QDarkStyle==3.0.2
qstylizer==0.2.1
QtAwesome==1.1.1
qtconsole==5.2.1
QtPy==1.11.2
regex==2021.11.10
requests
rope==0.22.0
Rtree==0.9.7
ruamel-yaml-conda
SecretStorage==3.3.1
six
snakeviz==2.1.1
snowballstemmer==2.2.0
sortedcontainers==2.4.0
Sphinx==4.3.1
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
spyder==5.2.0
spyder-kernels==2.2.0
tblib==1.7.0
testpath==0.5.0
text-unidecode==1.3
textdistance==4.2.2
three-merge==0.1.1
tinycss2==1.1.1
toml==0.10.2
tomli==1.2.2
tornado==6.1
tqdm
traitlets==5.1.1
typeguard==2.13.0
typing-extensions==4.0.0
ujson==4.3.0
urllib3
watchdog==2.1.6
wcwidth==0.2.5
webencodings==0.5.1
wrapt==1.13.3
wurlitzer==3.0.2
yapf==0.31.0
zipp==3.6.0
